Split implementation into more files
