+++ packages/node/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@powersync/node",
-  "version": "1.16.1",
+  "version": "0.1.0",
   "publishConfig": {
     "registry": "https://registry.npmjs.org/",
     "access": "public"
